backend/pyproject.toml
[project]
name = "multi-agent-task-solver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "langchain-tavily>=0.2.11",
    "langchain[anthropic,google-genai]>=0.3.27",
    "langgraph>=0.6.7",
    "matplotlib>=3.10.6",
    "pandas>=2.3.2",
    "python-dotenv>=1.0.0",
]
